Refactored progress bar update logic to ensure incremental updates during simulation.
- Updated `SolverEngine` to call `ReportProgress` more frequently and fixed threshold logic.
- Adjusted `MainViewModel` to handle progress updates correctly by converting values to percentages.
- Resolved ambiguity in `Math.Round` calls by explicitly casting arguments to `double`.
- Verified functionality without lowering `ProgressUpdateThresholdPercent`.

This resolves the issue of the progress bar jumping from 0 to 100% at the end of the simulation.

+++ NQueen.Domain/Interfaces/ISolverFrontEnd.cs
@@ -4,7 +4,7 @@ public interface ISolverFrontEnd
 {
     int DelayInMilliseconds { get; set; }
 
-    double ProgressValue { get; set; }
+    int ProgressValue { get; set; }
 
     event EventHandler<QueenPlacedEventArgs> QueenPlaced;
     event EventHandler<SolutionFoundEventArgs> SolutionFound;
